moved fetch_slots to my_modules, cause it's used as one of htcrystalball
