+++ my_modules/fetch_slots.py
